WIP: Added restart and logging to the VirtIO as a service
When DEBUG macro is defined in virtioUioTest.c and according
macros *_DBG are enabled in the virtioHostLib.c virtioHostNet.c
virtioLib.c and virtioVsm.c files, the debug and error messages
are logged using printf() call.

If the aforementioned macros are undefined in those files, error
messages are logged using syslog.

The VirtIO made work correctly with systemd service restart.
For this the VSM does device reset and the service is configured
to deliver SIGTERM signal and SIGKILL after a timeout.

Signed-off-by: Dmitriy Korovkin <Dmitriy.Korovkin@windriver.com>