VirtIO networking device throughput improvement
Issue: LINUXEXEC-30645

The throughput enhancement is achieved through the following
measures:

For the transmit side, the Generic segmentation offload (GSO) and
TCP segmentation offload (TSO) enabled. This also made the VNET header
required and allowed the removal of the functions that cut it
out of the packets.

For the receive side, virtioHostNetRxHandle() has been rearranged
to make sure that function handles the packets received from the TAP
interface and only if there are no any waits for the new packet
event.

For the receive side, increase the MTU to force the VirtIO device
driver providing more VirtIO buffers of larger size to handle the
larger receiving traffic.

Signed-off-by: Dmitriy Korovkin <Dmitriy.Korovkin@windriver.com>