virtio-uio: creata a separate process for BE device drivers
Issue: LINUXEXEC-32751

Added a way to allocate a separate process for a designated
BE device.

Introduced a new flag member in the virtioHostDrvInfo structure
if it is VIRTIO_HOST_FLAG_THREAD, device driver uses threads,
if it is VIRTIO_HOST_FLAG_PROCESS, the VSM does fork() before
creating a device.

When a BE device driver starts as a process, the VSM process
leaves the last of available cores to it and runs on the rest.

Signed-off-by: Dmitriy Korovkin <Dmitriy.Korovkin@windriver.com>