virtio-uio: Integrating vsock backend driver
Issue: LINUXEXEC-31284

The vsock driver aims to provide full virtio-vsock support to software
running inside the guest VM, while bypassing vhost kernel code on the host.

It implements the vsock device model, and and mediates communication between
AF_UNIX sockets (on the host end) and AF_VSOCK sockets (on the guest end).

In order to provide channel multiplexing the guest AF_VSOCK ports are mapped
1:1 to AF_UNIX sockets on the host. The virtio-vsock host driver will create
a AF_UNIX socket on for per guest (/tmp/host.${CID})

When host want to connect to a guest, on host side, the application should
open /tmp/host.${CID}, and send a connection command, full process as below:

1. Guest: create an AF_VSOCK socket and listen() on <port_num>
2. Host: connect() to AF_UNIX e.g /tmp/host.16
3. Host: send() "CONNECT <port_num>\n"
4. Guest: accept() the new connection.
5. Host: read() "OK <assigned_hostside_port>\n".

When host detected a connection request from guest(a VIRTIO_VSOCK_OP_REQUEST
packet), it tries to forward the connection to an AF_UNIX socket listening on
the host (/tmp/vsock.${CID}_${PORT}). If no such socket exists, or no one is
listening on it, a connection cannot be established, and a VIRTIO_VSOCK_OP_RST
packet will be sent back to the guest, full process as below:
1. Host: create and listen on an AF_UNIX socket at (e.g. /tmp/vsock.16_1234)
2. Guest: create an AF_VSOCK socket and connect to HOST_CID (integer value 2)
and PORT;
3. Host: accept() the new connection.

The above process is full compatible with Amazon Firecracker so the OSS package
which support Amazon Firecracker's vsock can be used here directly.

Signed-off-by: Xiangyu Chen <xiangyu.chen@windriver.com>