virtio-console: modified BE device driver for process model
Issue: LINUXEXEC-32751

Mevent initialization is moved to device creation stage for the process
so that mevent thread is created after the BE device is actually being
forked. In this case the initialization needs to wait for the Mevent
handling thread to start. For the thread model the Mevent thread is
one for all the devices and starts on the BE device driver
initialization.

Removed the dispatch thread and dispatch queues as they are functionally
replaced by POSIX message queues that hold the requests for devices.

Added syslog output for errors to be inline with the rest of the
BE device drivers.

Signed-off-by: Dmitriy Korovkin <Dmitriy.Korovkin@windriver.com>