virtio-uio: request queues processing improvement
Issue: LINUXEXEC-32751

Added hecking if the VirtIO queue contains any new buffer before
releasing the semaphore that holds the thread that processes
the queue data. This allows to avoid the unneeded context
switch to the thread which will only check that  the queue
contains no new buffers and go to sleep again.

Signed-off-by: Dmitriy Korovkin <Dmitriy.Korovkin@windriver.com>